api/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
torch
transformers
oracledb
python-multipart
pydantic